refactor(mapgen): close the recipe step kind
Require each recipe step leaf to own config.ts and step.ts, with an optional attached viz.ts facet. Inline step-private execution helpers, relocate reusable policy and telemetry to qualified owners, retire the stale placement-specific shape rule, and preserve focused placement, ecology, hydrology, and morphology behavior.
