+++ .habitat/blueprints/recipe-step/require_recipe_step_runtime_owner_shape/rule.json
@@ -12,16 +12,14 @@
   },
   "ownerProject": "mod-swooper-maps",
   "lane": "enforced",
-  "forbids": "recipe step step.ts modules without one semantically named *Step = createStep(config, ...) executable, with additional runtime exports or re-exports, with a default export, or with runtime callbacks that confuse the admitted step value with the imported static config owner",
-  "why": "A recipe step runtime owner imports its literal local config contract and exposes one semantically named *Step executable as its public runtime identity. Runtime callbacks receive the admitted value as stepConfig (or _stepConfig when unused), so static config.id metadata and executable configuration cannot silently shadow one another. Local helpers and erased type/interface exports remain available without creating competing exported runtime surfaces. Exact step.ts acquisition selects the kind, Core's createStep contract owns constructor admission, and Knip owns unused private construction.",
-  "remediate": "Import config from ./config.js, retain one documented <Name>Step = createStep(config, ...) export, name the admitted runtime callback value stepConfig or _stepConfig, keep non-authority helpers local, and delete default or additional runtime exports and re-exports.",
-  "message": "Recipe step step.ts must expose only its semantically named *Step executable over the local config contract, with runtime callbacks using the distinct stepConfig identity.",
+  "forbids": "recipe step step.ts modules without one semantically named *Step = createStep(config, ...) executable, with additional runtime exports or re-exports, with a default export, with runtime callbacks that confuse the admitted step value with the imported static config owner, or with a sibling viz import detached from the executable viz facet",
+  "why": "A recipe step runtime owner imports its literal local config contract and exposes one semantically named *Step executable as its public runtime identity. Runtime callbacks receive the admitted value as stepConfig (or _stepConfig when unused), so static config.id metadata and executable configuration cannot silently shadow one another. An optional sibling viz.ts is implementation placement for the executable's first-class viz facet, not a general helper cabinet. Local helpers and erased type/interface exports remain available without creating competing exported runtime surfaces. Exact step.ts acquisition selects the kind, Core's createStep contract owns constructor admission, and Knip owns unused private construction or orphan files.",
+  "remediate": "Import config from ./config.js, retain one documented <Name>Step = createStep(config, ...) export, name the admitted runtime callback value stepConfig or _stepConfig, attach any sibling ./viz.js import through the createStep viz facet, keep non-authority helpers local, and delete default or additional runtime exports and re-exports.",
+  "message": "Recipe step step.ts must expose only its semantically named *Step executable over the local config contract, use the distinct stepConfig runtime identity, and attach any sibling viz module through the executable viz facet.",
   "pathCoverage": [
     {
       "kind": "exact-path",
-      "patterns": [
-        "mods/*/src/recipes/*/stages/**/steps/*/step.ts"
-      ]
+      "patterns": ["mods/*/src/recipes/*/stages/**/steps/*/step.ts"]
     }
   ],
   "hookCheck": true,
@@ -36,9 +34,7 @@
     "patternName": "require_recipe_step_runtime_owner_shape",
     "acquisition": {
       "kind": "check",
-      "roots": [
-        "mods"
-      ]
+      "roots": ["mods"]
     }
   }
 }
