--- .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/stages/placement/rules/require_typed_placement_outcomes_before_apply/baseline.json
@@ -1,1 +1,0 @@
-[]
