+++ docs/system/libs/mapgen/policies/NORMALIZATION-GUARDRAILS.md
@@ -32,7 +32,7 @@ G1-G9 enforcement scope:
 | G5 sibling stage `steps/` imports | `normalize-ecology-topology` | Habitat `prohibit_sibling_stage_private_step_imports` | Standard stages do not import sibling stages' private `steps/` modules. |
 | G6 recipe doc stage drift | `normalize-ecology-topology` | No direct current rule. `verify_runtime_stage_order_matches_contract_manifest` checks runtime against `contract-manifest.ts`; `validate_mapgen_docs_anchors_and_references` checks MapGen doc anchors and reference policy. | No current registered rule proves that `STANDARD-RECIPE.md` stage order matches source; keep that source-doc comparison in review until a manifest owns it. |
 | G7 superseded current-stage ids | `normalize-ecology-topology`, `normalize-projection-lakes` | Generic domain, module, and recipe-stage topology blueprints protect current source kinds; there is no direct current doc-content rule. | No current registered rule decides whether evergreen prose presents a retired stage id as current source shape. |
-| G8 hidden placement sub-concerns | `normalize-placement-contracts`, `normalize-placement-reconciliation` | Habitat `require_typed_placement_outcomes_before_apply` | Placement apply does not directly call official resource/discovery generators and typed outcomes exist before apply. |
+| G8 hidden placement sub-concerns | `normalize-placement-contracts`, `normalize-placement-reconciliation` | Generic recipe-step topology plus typed artifact contracts and direct Step behavior tests. The retired exact-file rule no longer names a deleted `placement/apply.ts` helper. | Resource and discovery materialization remain explicit executable steps over typed plans and outcomes rather than hidden apply sub-concerns. |
 | G9 wrapper-only `advanced` stage config | `normalize-config-surface` | Habitat `prohibit_wrapper_only_advanced_config` | Standard recipe source and map configs do not reintroduce persisted SDK-native `advanced` wrappers. |
 | G10 visualization contract owner surfaces | `normalize-viz-contract-owners` | Habitat `require_shared_visualization_contracts_at_stage_surfaces` | No collection-wide `steps/viz.ts` hub exists. Recipe-root styles and stage/step projection ownership remain normative review and import-boundary concerns. |
 | G11 SDK map runtime entrypoint boundary | `normalize-sdk-mapgen-runtime-entrypoint` | Habitat `require_explicit_mapgen_sdk_opt_in` | The SDK root remains runtime-neutral; Civ7 map runtime helpers are exposed only from the MapGen subpath. |
@@ -50,7 +50,7 @@ completed implementation record.
 
 ## Commands
 
-- `bun habitat check --rule prohibit_milestone_prefixed_standard_recipe_tag_catalog_names --rule prohibit_domain_artifacts_modules --rule preserve_mapgen_core_runtime_neutrality --rule require_public_domain_surfaces_in_recipes_and_maps --rule prohibit_sibling_stage_private_step_imports --rule require_typed_placement_outcomes_before_apply --rule prohibit_wrapper_only_advanced_config --rule require_shared_visualization_contracts_at_stage_surfaces --rule require_explicit_mapgen_sdk_opt_in`
+- `bun habitat check --rule prohibit_milestone_prefixed_standard_recipe_tag_catalog_names --rule prohibit_domain_artifacts_modules --rule preserve_mapgen_core_runtime_neutrality --rule require_public_domain_surfaces_in_recipes_and_maps --rule prohibit_sibling_stage_private_step_imports --rule require_recipe_step_source_topology --rule prohibit_wrapper_only_advanced_config --rule require_shared_visualization_contracts_at_stage_surfaces --rule require_explicit_mapgen_sdk_opt_in`
 - `bun habitat check --rule validate_mapgen_docs_anchors_and_references`
 
 `habitat check` selects bounded rules with repeatable `--rule`; it has no
