--- mods/mod-swooper-maps/src/recipes/standard/stages/ecology/projection/steps/features-apply/apply.ts
@@ -1,35 +1,0 @@
-import type { FeatureKey } from "@civ7/map-policy";
-import type { FeatureIntentKey } from "@mapgen/domain/ecology/modules/features/model/atoms/index.js";
-
-const FEATURE_KEY_BY_INTENT: Readonly<Record<FeatureIntentKey, FeatureKey>> = {
-  forest: "FEATURE_FOREST",
-  rainforest: "FEATURE_RAINFOREST",
-  taiga: "FEATURE_TAIGA",
-  "savanna-woodland": "FEATURE_SAVANNA_WOODLAND",
-  "sagebrush-steppe": "FEATURE_SAGEBRUSH_STEPPE",
-  marsh: "FEATURE_MARSH",
-  "tundra-bog": "FEATURE_TUNDRA_BOG",
-  mangrove: "FEATURE_MANGROVE",
-  oasis: "FEATURE_OASIS",
-  "watering-hole": "FEATURE_WATERING_HOLE",
-  reef: "FEATURE_REEF",
-  "cold-reef": "FEATURE_COLD_REEF",
-  atoll: "FEATURE_ATOLL",
-  lotus: "FEATURE_LOTUS",
-  ice: "FEATURE_ICE",
-  "desert-floodplain-minor": "FEATURE_DESERT_FLOODPLAIN_MINOR",
-  "desert-floodplain-navigable": "FEATURE_DESERT_FLOODPLAIN_NAVIGABLE",
-  "grassland-floodplain-minor": "FEATURE_GRASSLAND_FLOODPLAIN_MINOR",
-  "grassland-floodplain-navigable": "FEATURE_GRASSLAND_FLOODPLAIN_NAVIGABLE",
-  "plains-floodplain-minor": "FEATURE_PLAINS_FLOODPLAIN_MINOR",
-  "plains-floodplain-navigable": "FEATURE_PLAINS_FLOODPLAIN_NAVIGABLE",
-  "tropical-floodplain-minor": "FEATURE_TROPICAL_FLOODPLAIN_MINOR",
-  "tropical-floodplain-navigable": "FEATURE_TROPICAL_FLOODPLAIN_NAVIGABLE",
-  "tundra-floodplain-minor": "FEATURE_TUNDRA_FLOODPLAIN_MINOR",
-  "tundra-floodplain-navigable": "FEATURE_TUNDRA_FLOODPLAIN_NAVIGABLE",
-};
-
-/** Resolves an admitted Ecology feature intent to its official Civ7 projection identity. */
-export function resolveFeatureKeyForIntent(intent: FeatureIntentKey): FeatureKey {
-  return FEATURE_KEY_BY_INTENT[intent];
-}
