--- mods/mod-swooper-maps/src/recipes/standard/stages/ecology/projection/steps/plot-biomes/helpers/apply.ts
@@ -1,7 +1,0 @@
-/**
- * Clamps a value to a 0..255 integer for engine byte fields.
- */
-export function clampToByte(value: number): number {
-  if (!Number.isFinite(value)) return 0;
-  return Math.max(0, Math.min(255, Math.round(value)));
-}
