+++ mods/mod-swooper-maps/test/recipes/swooper-physics-standard/stages/placement/steps/assign-starts.materialization.test.ts
@@ -1,67 +1,53 @@
 import { describe, expect, it } from "bun:test";
+
 import { createMockAdapter } from "@civ7/adapter";
-import placementDomain from "@mapgen/domain/placement/router";
-import { admitMapSetup, createMapContext } from "@swooper/mapgen-core";
-import type { Static } from "@swooper/mapgen-core/authoring";
-import { hexDistanceOddQPeriodicX } from "@swooper/mapgen-core/lib/grid";
-import { runAdmittedOperationForTest } from "@swooper/mapgen-core/testing";
+import { artifacts as pedologyArtifacts } from "@mapgen/domain/ecology/modules/pedology/artifacts/index.js";
+import { artifacts as climateArtifacts } from "@mapgen/domain/hydrology/modules/climate/artifacts/index.js";
+import { artifacts as hydrographyArtifacts } from "@mapgen/domain/hydrology/modules/hydrography/artifacts/index.js";
+import { artifacts as morphologyLandformsArtifacts } from "@mapgen/domain/morphology/modules/landforms/artifacts/index.js";
+import { artifacts as morphologyShelfArtifacts } from "@mapgen/domain/morphology/modules/shelf/artifacts/index.js";
+import { artifacts as placementRegionArtifacts } from "@mapgen/domain/placement/modules/regions/artifacts/index.js";
 import { artifacts as placementStartArtifacts } from "@mapgen/domain/placement/modules/starts/artifacts/index.js";
-import { materializeStartAssignment } from "../../../../../../src/recipes/standard/stages/placement/steps/assign-starts/materialize.js";
+import { artifacts as placementWonderArtifacts } from "@mapgen/domain/placement/modules/wonders/artifacts/index.js";
+import placement from "@mapgen/domain/placement/router";
+import { artifacts as resourceSiteArtifacts } from "@mapgen/domain/resources/modules/sites/artifacts/index.js";
+import { admitMapSetup, createMapContext, type MapContext } from "@swooper/mapgen-core";
+import {
+  readValidatedArtifact,
+  type Static,
+  type StepRuntimeOps,
+} from "@swooper/mapgen-core/authoring";
+import {
+  buildStepTestDependencies,
+  normalizeOperationSelectionForTest,
+  publishTestArtifact,
+  withMapContextExecutionForTest,
+} from "@swooper/mapgen-core/testing";
+
+import { AssignStartsStep } from "../../../../../../src/recipes/standard/stages/placement/steps/assign-starts/step.js";
 import { TEST_MAP_SEED, TEST_MAP_SIZE } from "../../../../../setup.js";
 
-const { planStarts } = placementDomain.starts.ops;
-type StartInput = Static<(typeof planStarts)["input"]>;
+type AssignStartsConfig = Static<(typeof AssignStartsStep.contract)["schema"]>;
+type AssignStartsOps = StepRuntimeOps<NonNullable<(typeof AssignStartsStep.contract)["ops"]>>;
+type LandTile = readonly [x: number, y: number];
 
-function makeInput(
-  width: number,
-  height: number,
-  playersLandmass1: number,
-  playersLandmass2: number,
-  alivePlayerIds?: number[]
-) {
-  const size = width * height;
-  const landmassTileCounts: number[] = [];
-  return {
-    baseStarts: { playersLandmass1, playersLandmass2 },
-    ...(alivePlayerIds === undefined ? {} : { alivePlayerIds }),
-    width,
-    height,
-    landMask: new Uint8Array(size),
-    slotByTile: new Uint8Array(size),
-    landmassIdByTile: new Int32Array(size).fill(-1),
-    landmassTileCounts,
-    coastalLand: new Uint8Array(size),
-    distanceToCoast: new Uint16Array(size),
-    shelfMask: new Uint8Array(size),
-    elevation: new Int16Array(size),
-    fertility: new Float32Array(size).fill(0.55),
-    effectiveMoisture: new Float32Array(size).fill(0.55),
-    surfaceTemperature: new Float32Array(size).fill(16),
-    aridityIndex: new Float32Array(size).fill(0.35),
-    riverClass: new Uint8Array(size),
-    lakeMask: new Uint8Array(size),
-  } satisfies StartInput;
-}
-
-function addLandmass(
-  input: ReturnType<typeof makeInput>,
-  tiles: ReadonlyArray<readonly [number, number]>
-): void {
-  input.landmassTileCounts[0] = tiles.length;
-  for (const [x, y] of tiles) {
-    const plotIndex = y * input.width + x;
-    input.landMask[plotIndex] = 1;
-    input.slotByTile[plotIndex] = 1;
-    input.landmassIdByTile[plotIndex] = 0;
-    input.coastalLand[plotIndex] = 1;
+const ASSIGN_STARTS_OP_CONTRACTS = AssignStartsStep.contract.ops!;
+const STARTS_RUNTIME_OP = Object.assign(
+  (
+    input: Parameters<AssignStartsOps["starts"]>[0],
+    config: Parameters<AssignStartsOps["starts"]>[1]
+  ): ReturnType<AssignStartsOps["starts"]> => placement.starts.ops.planStarts.run(input, config),
+  {
+    id: ASSIGN_STARTS_OP_CONTRACTS.starts.id,
+    kind: ASSIGN_STARTS_OP_CONTRACTS.starts.kind,
   }
-}
+);
+const ASSIGN_STARTS_OPS = { starts: STARTS_RUNTIME_OP } satisfies AssignStartsOps;
 
-function plan(
-  input: StartInput,
-  configure?: (config: (typeof planStarts.defaultConfig)["config"]) => void
-) {
-  const selection = structuredClone(planStarts.defaultConfig);
+function assignStartsConfig(
+  configure?: (config: (typeof placement.starts.ops.planStarts.defaultConfig)["config"]) => void
+): AssignStartsConfig {
+  const selection = structuredClone(placement.starts.ops.planStarts.defaultConfig);
   selection.config.minContiguousLandTiles = 12;
   selection.config.minExpansionLandTiles = 6;
   selection.config.minIslandClusterLandTiles = 8;
@@ -69,15 +55,19 @@ function plan(
   selection.config.spacingFloorTiles = 2;
   selection.config.desiredSpacingTiles = 4;
   configure?.(selection.config);
-  return runAdmittedOperationForTest(planStarts, input, selection);
+  return {
+    starts: normalizeOperationSelectionForTest(placement.starts.ops.planStarts, selection),
+  };
 }
 
-function contextFor() {
+function createAssignStartsContext(alivePlayerIds: readonly number[]) {
   const adapter = createMockAdapter({
     ...TEST_MAP_SIZE.dimensions,
     mapInfo: TEST_MAP_SIZE.mapInfo,
     mapSizeId: TEST_MAP_SIZE.id,
+    aliveMajorCount: alivePlayerIds.length,
   });
+  adapter.getAliveMajorIds = () => [...alivePlayerIds];
   const context = createMapContext({
     setup: admitMapSetup({
       mapSeed: TEST_MAP_SEED,
@@ -92,75 +82,209 @@ function contextFor() {
   return { adapter, context };
 }
 
-describe("start materializer (thin shell)", () => {
-  it("stamps every seated intent via setStartPosition with the op's playerId and validates", () => {
-    const { width, height } = TEST_MAP_SIZE.dimensions;
-    const input = makeInput(width, height, 2, 0, [4, 9]);
-    addLandmass(
-      input,
-      Array.from({ length: 80 }, (_value, i) => [1 + (i % 10), 1 + Math.floor(i / 10)] as const)
+function publishAssignStartsInputs(context: MapContext, landTiles: readonly LandTile[]): void {
+  const { width, height } = context.setup.dimensions;
+  const size = width * height;
+  const landMask = new Uint8Array(size);
+  const landmassIdByTile = new Int32Array(size).fill(-1);
+  const slotByTile = new Uint8Array(size);
+  const coastalLand = new Uint8Array(size);
+  for (const [x, y] of landTiles) {
+    const plotIndex = y * width + x;
+    landMask[plotIndex] = 1;
+    landmassIdByTile[plotIndex] = 0;
+    slotByTile[plotIndex] = 1;
+    coastalLand[plotIndex] = 1;
+  }
+  const xs = landTiles.map(([x]) => x);
+  const ys = landTiles.map(([, y]) => y);
+  const landmasses =
+    landTiles.length === 0
+      ? []
+      : [
+          {
+            id: 0,
+            tileCount: landTiles.length,
+            coastlineLength: 0,
+            bbox: {
+              west: Math.min(...xs),
+              east: Math.max(...xs),
+              south: Math.min(...ys),
+              north: Math.max(...ys),
+            },
+          },
+        ];
+
+  publishTestArtifact(context, resourceSiteArtifacts.resourcePlan, {
+    width,
+    height,
+    seed: TEST_MAP_SEED,
+    plannedCount: 0,
+    rotationCount: 0,
+    rangeFloorCount: 0,
+    regionMinimumCount: 0,
+    siteSpacingTiles: 0,
+    equitySkippedSiteCount: 0,
+    intents: [],
+    perType: [],
+    regionMinimums: [],
+    settings: {
+      density: 1,
+      sparsity: 0,
+      rarityFidelity: 1,
+      perTypeSpacingFloorScale: 1,
+      equityMaxDensityRatio: 1,
+      affinityRuleCount: 0,
+      affinityRules: [],
+    },
+  });
+  publishTestArtifact(context, placementWonderArtifacts.naturalWonderPlacement, {
+    plannedCount: 0,
+    targetCount: 0,
+    placedCount: 0,
+    terrainAdjustedCount: 0,
+    skippedOutOfBoundsCount: 0,
+    rejectedCount: 0,
+    shortfallCount: 0,
+    rejectionExamples: [],
+    coordinateEvidence: {
+      version: 1,
+      placed: { count: 0, hash32: "811c9dc5" },
+      rejected: { count: 0, hash32: "811c9dc5" },
+    },
+    coordinateRows: [],
+    observedNaturalWonderPlotIndices: [],
+  });
+  publishTestArtifact(context, placementRegionArtifacts.landmassRegionSlotByTile, {
+    slotByTile,
+  });
+  publishTestArtifact(context, morphologyLandformsArtifacts.topography, {
+    elevation: new Int16Array(size).fill(500),
+    seaLevel: 0,
+    landMask,
+    bathymetry: new Int16Array(size),
+  });
+  publishTestArtifact(context, morphologyLandformsArtifacts.landmasses, {
+    landmasses,
+    landmassIdByTile,
+  });
+  publishTestArtifact(context, morphologyLandformsArtifacts.mountains, {
+    mountainMask: new Uint8Array(size),
+    mountainRegionMask: new Uint8Array(size),
+    mountainRegionIdByTile: new Int32Array(size).fill(-1),
+    hillMask: new Uint8Array(size),
+    foothillMask: new Uint8Array(size),
+    roughLandMask: new Uint8Array(size),
+    orogenyPotential: new Uint8Array(size),
+    fracturePotential: new Uint8Array(size),
+    roughnessPotential: new Uint8Array(size),
+  });
+  publishTestArtifact(context, morphologyLandformsArtifacts.volcanoes, {
+    volcanoMask: new Uint8Array(size),
+    volcanoes: [],
+  });
+  publishTestArtifact(context, morphologyShelfArtifacts.shelf, {
+    shelfMask: new Uint8Array(size),
+    coastalLand,
+    coastalWater: new Uint8Array(size),
+    distanceToCoast: new Uint16Array(size),
+  });
+  publishTestArtifact(context, hydrographyArtifacts.hydrography, {
+    runoff: new Float32Array(size),
+    discharge: new Float32Array(size),
+    riverClass: new Uint8Array(size),
+    flowDir: new Int32Array(size).fill(-1),
+    sinkMask: new Uint8Array(size),
+    outletMask: new Uint8Array(size),
+    basinId: new Int32Array(size).fill(-1),
+    routingElevation: new Float32Array(size),
+    depressionDepth: new Float32Array(size),
+    terminalType: new Uint8Array(size),
+  });
+  publishTestArtifact(context, hydrographyArtifacts.lakePlan, {
+    width,
+    height,
+    lakeMask: new Uint8Array(size),
+    plannedLakeTileCount: 0,
+    sinkLakeCount: 0,
+  });
+  publishTestArtifact(context, climateArtifacts.climateIndices, {
+    surfaceTemperatureC: new Float32Array(size).fill(16),
+    effectiveMoisture: new Float32Array(size).fill(0.5),
+    pet: new Float32Array(size),
+    aridityIndex: new Float32Array(size).fill(0.5),
+    freezeIndex: new Float32Array(size),
+  });
+  publishTestArtifact(context, pedologyArtifacts.pedology, {
+    width,
+    height,
+    soilType: new Uint8Array(size),
+    fertility: new Float32Array(size).fill(0.5),
+  });
+}
+
+function runAssignStartsStep(
+  context: MapContext,
+  landTiles: readonly LandTile[],
+  config: AssignStartsConfig = assignStartsConfig()
+): void {
+  withMapContextExecutionForTest(context, (stepContext) => {
+    publishAssignStartsInputs(stepContext, landTiles);
+    AssignStartsStep.run(
+      stepContext,
+      config,
+      ASSIGN_STARTS_OPS,
+      buildStepTestDependencies(AssignStartsStep, stepContext)
     );
-    const planned = plan(input);
-    const { adapter, context } = contextFor();
+  });
+}
 
-    const assignment = materializeStartAssignment({
-      context,
-      plan: planned,
-      setStartPosition: adapter.setStartPosition.bind(adapter),
-    });
+describe("assign starts step", () => {
+  it("stamps every planned seat with the operation-owned player and plot identities", () => {
+    const landTiles = Array.from(
+      { length: 80 },
+      (_value, index) => [1 + (index % 10), 1 + Math.floor(index / 10)] as const
+    );
+    const { adapter, context } = createAssignStartsContext([4, 9]);
+
+    runAssignStartsStep(context, landTiles);
+    const assignment = readValidatedArtifact(context, placementStartArtifacts.startAssignment);
 
     expect(assignment.assigned).toBe(2);
-    expect(adapter.calls.setStartPosition.map((call) => call.playerId).sort()).toEqual([4, 9]);
+    expect(adapter.calls.setStartPosition.map(({ playerId }) => playerId).sort()).toEqual([4, 9]);
     expect(
-      adapter.calls.setStartPosition.map((call) => call.plotIndex).sort((a, b) => a - b)
+      adapter.calls.setStartPosition.map(({ plotIndex }) => plotIndex).sort((a, b) => a - b)
     ).toEqual([...assignment.positions].sort((a, b) => a - b));
-    expect(placementStartArtifacts.startAssignment.validate(assignment)).toEqual([]);
-    const [firstPosition, secondPosition] = assignment.positions;
-    if (firstPosition === undefined || secondPosition === undefined) {
-      throw new Error("Expected two materialized start positions.");
-    }
-    const spacing = hexDistanceOddQPeriodicX(firstPosition, secondPosition, input.width);
-    expect(spacing).toBeGreaterThanOrEqual(2);
   });
 
-  it("hard-fails ONLY when literally zero settleable land candidates exist", () => {
-    const { width, height } = TEST_MAP_SIZE.dimensions;
-    const input = makeInput(width, height, 1, 0);
-    const planned = plan(input);
-    const { adapter, context } = contextFor();
+  it("hard-fails only when requested seats have no settleable land candidate", () => {
+    const { context } = createAssignStartsContext([4]);
 
-    expect(() =>
-      materializeStartAssignment({
-        context,
-        plan: planned,
-        setStartPosition: adapter.setStartPosition.bind(adapter),
-      })
-    ).toThrow(/No settleable land candidates/);
+    expect(() => runAssignStartsStep(context, [])).toThrow(/No settleable land candidates/);
   });
 
-  it("materializes degraded plans as data (no assign-or-throw)", () => {
-    const { width, height } = TEST_MAP_SIZE.dimensions;
-    const input = makeInput(width, height, 3, 0);
-    addLandmass(input, [
-      [2, 2],
-      [5, 4],
-    ]);
-    const planned = plan(input, (config) => {
-      config.spacingFloorTiles = 1;
-      config.desiredSpacingTiles = 2;
+  it("publishes unseated players as degraded assignment data", () => {
+    const { adapter, context } = createAssignStartsContext([4, 9, 11]);
+    const config = assignStartsConfig((selection) => {
+      selection.spacingFloorTiles = 1;
+      selection.desiredSpacingTiles = 2;
     });
-    const { adapter, context } = contextFor();
 
-    const assignment = materializeStartAssignment({
+    runAssignStartsStep(
       context,
-      plan: planned,
-      setStartPosition: adapter.setStartPosition.bind(adapter),
-    });
+      [
+        [2, 2],
+        [5, 4],
+      ],
+      config
+    );
+    const assignment = readValidatedArtifact(context, placementStartArtifacts.startAssignment);
 
-    expect(assignment.assigned).toBe(2);
-    expect(assignment.unseatedCount).toBe(1);
-    expect(assignment.status).toBe("degraded");
-    expect(adapter.calls.setStartPosition.length).toBe(2);
-    expect(placementStartArtifacts.startAssignment.validate(assignment)).toEqual([]);
+    expect(assignment).toMatchObject({
+      assigned: 2,
+      unseatedCount: 1,
+      status: "degraded",
+    });
+    expect(adapter.calls.setStartPosition).toHaveLength(2);
   });
 });
